Generalized pipeline
File "Find_lane_lines.py" created with all the important parts of the
pipeline. The file contains the following functions:
- grayscale: Converts the image to gray scale
- gaussian_blur: Filters the image with gaussian noise kernel
- canny: Applies Canny transform to the image
- region_of_interest: Selects the part of the edges image where the
lines are located and deletes the rest.
- hough_lines: Uses the Hough transformation to determine points
defining lines. This correspond to the real lines in the image, so if a
line is dashed, it is not recognized as an entire line. To recognize the
entire lines the points are separated in two groups, one with positive
slope and other one with negative slope. and with those points the
position X at the bottom of the image and at the far horizon of the
image are calculated and with them the lines are drawn in a new image.
This new image only contains the detected lines.
- weighted_img: it overlaps the image with the detected lines to the
original image.

 Pipeline tested with image "imgWhite2".