Pipeline tested with video 1
Pipeline tested and working with video 1 "solidWhiteRight.mp4".

The function "hough_lines" on the file "Find_lane_lines.py" had to be
modified. Now it is everywhere checked, if it can result a division by
0. Also it is checked if the arrays including the X values are empty or
not, because if they are empty the average will result in a NaN. Instead
the last average values (of the last execution of the function) are used
if the arrays are empty. To save those values, static variables are used
(actually attributes of the function).