Pipeline tested with video 2 and with challenge video
The pipeline is now working well with the video 2 and also with the
challenge.

With the video 2 it worked right away without doing modifications, but
with the challenge video the following modifications needed to be done
so it works:
- Function to get the vertices of the region of interest (houghVertices)
changed, because the video has another resolution and the car can be seen at
the bottom of the image. Now the bottom of the region of interest is also
variable.

- Function to detect the lines (hough_lines) changed, so only the points
within a range are used to draw the lines. So could the accuracy of the
line positions be increased dramatically. Also it was added an exception
for the case that the OpenCV function "HoughLinesP" does not find any
line. In that case the last detected line is used for the cycle.

- Parameters for the challenge video optimized:
-- Canny_low_threshold from 50 to 20
-- Hough_theta from 2° to 1°
-- Hough_min_line_len from 10 to 40
-- Hough_max_line_gap from 10 to 2